fixed some schematic errors, created a board layout for testing
Well, the schematic was fixed because the Tx and Rx were mixed up (the cellular module's Tx needs to go to the iPod's Tx--yeah, I found that weird, too). I also added connections to the iPod's 3.3V output and added a level-shifting resistor to the iPod's Tx line so as to not fry the cellular module's 2.8V CMOS level serial circuitry.